requirements.txt
fastapi==0.114.0
uvicorn==0.23.2
sqlalchemy==2.0.34
databases==0.9.0
python-dotenv==1.0.1
aiomysql==0.0.21
pydantic==1.9.0
openai==1.47.1
pandas==2.2.2
google-cloud-aiplatform==1.63.0